Change of method definition
Change from inline to externally, because otherwise there might be heaps
of resources used when thousands of notes are loaded.
Changed the name of the class, because ~~f*ck Hungarian Notation~~
Hungarian Notation is outdated, from official words, and there isn't a
created class in C++ called Note or Chart, so I'll do so.
(Hope no one see this)