Change of the actualization of struct Note
It is shown that Cytus II uses tick to measure the time. I wanted struct
Note to be capable of handling both Cytus' time (as in double) and Cytus
II's tick (as in int), so I rewrote it into a template.
To utilize the vector type in C++, I overloaded the operator < to make
sure it can be sorted by time. Also I changed the note list to a vector
of struct Note.